refactor: remove suwa-sh/ddd dependency
- Create local ProcessStatus enum in infra.classification
- Create local Input/Output interfaces in infra.policy
- Refactor BaseLayerSuperType to standalone abstract class
  (remove GenericLayerSuperType/LayerSuperType inheritance)
- Remove Facade interface from BaseFacade
- Replace DddMessageSource/DddDdSource parent references with null
- Update all imports from me.suwash.ddd.* to local types
- Remove ddd dependency and version property from pom.xml
