+++ pom.xml
@@ -25,7 +25,6 @@
         <dirname.assembly.work>compare_files_${project.version}</dirname.assembly.work>
         <dir.assembly.work>${project.build.directory}/${dirname.assembly.work}</dir.assembly.work>
 
-        <version.ddd>1.0.0</version.ddd>
         <version.test>1.0.1</version.test>
     </properties>
 
@@ -256,18 +255,6 @@
                 <exclusion>
                     <groupId>org.apache.commons</groupId>
                     <artifactId>commons-lang3</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>me.suwash.commons</groupId>
-            <artifactId>ddd</artifactId>
-            <version>${version.ddd}</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.apache.velocity</groupId>
-                    <artifactId>velocity</artifactId>
                 </exclusion>
             </exclusions>
         </dependency>
