+++ src/main/java/me/suwash/tools/comparefiles/infra/classification/ProcessStatus.java
@@ -1,0 +1,15 @@
+package me.suwash.tools.comparefiles.infra.classification;
+
+/**
+ * 処理ステータス。
+ */
+public enum ProcessStatus {
+    /** 処理中。 */
+    Processing,
+    /** 成功。 */
+    Success,
+    /** 警告。 */
+    Warning,
+    /** 失敗。 */
+    Failure
+}
