+++ src/main/java/me/suwash/tools/comparefiles/infra/policy/Input.java
@@ -1,0 +1,7 @@
+package me.suwash.tools.comparefiles.infra.policy;
+
+/**
+ * 入力データモデルマーカーインタフェース。
+ */
+public interface Input {
+}
