+++ src/main/java/me/suwash/tools/comparefiles/main/CompareFileRegex.java
@@ -3,7 +3,7 @@ package me.suwash.tools.comparefiles.main;
 import java.io.File;
 import java.util.List;
 
-import me.suwash.ddd.classification.ProcessStatus;
+import me.suwash.tools.comparefiles.infra.classification.ProcessStatus;
 import me.suwash.tools.comparefiles.ap.FileRegexCompareFacade;
 import me.suwash.tools.comparefiles.ap.dto.FileRegexCompareFacadeInput;
 import me.suwash.tools.comparefiles.ap.dto.FileRegexCompareFacadeOutput;
